requirements.txt
tensorflow>=2.16
gradio>=4.0
numpy
pillow